Baselines/RLPrompt/requirements.txt
torch
numpy
pandas
typing
wandb
transformers
hydra-core>=1.3.0
matplotlib
seaborn
tqdm
pyyaml